Move audio_track.dart to schemas
