+++ lib/audiobookshelf_api.dart
@@ -1,15 +1,15 @@
-export 'src/audiobookshelf_api_base.dart';
-export 'src/login_response.dart';
-export 'src/user.dart';
 export 'src/audiobook_progress.dart';
+export 'src/audiobook_search_result.dart';
+export 'src/audiobookshelf_api_base.dart';
+export 'src/collection.dart';
+export 'src/login_response.dart';
+export 'src/login_response.dart';
+export 'src/models/schemas/audio_track.dart';
 export 'src/models/schemas/library.dart';
 export 'src/models/schemas/library_item.dart';
 export 'src/models/schemas/media.dart';
 export 'src/models/schemas/series.dart';
-export 'src/audiobook_search_result.dart';
-export 'src/login_response.dart';
 export 'src/search_response.dart';
 export 'src/series_search_result.dart';
-export 'src/audio_track.dart';
-export 'src/collection.dart';
+export 'src/user.dart';
 export 'src/utils.dart';
