+++ lib/src/models/schemas/audio_track.dart
